+++ webulous-mvc/src/main/java/uk/ac/ebi/spot/webulous/controller/RestrictionController.java
@@ -1,6 +1,7 @@
 package uk.ac.ebi.spot.webulous.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
@@ -26,6 +27,9 @@ import java.util.List;
 @RequestMapping("/restrictions")
 public class RestrictionController {
 
+    @Value("${webulous.ui.readonly}")
+    boolean readOnly = false;
+
     @Autowired
     private RestrictionService restrictionService;
 
@@ -35,12 +39,18 @@ public class RestrictionController {
     }
 
     @RequestMapping(value = "", produces = MediaType.TEXT_HTML_VALUE, method = RequestMethod.GET)
-    public String showRestrictionRuns() {
+    public String showRestrictionRuns(Model model) {
+        if (readOnly) {
+            model.addAttribute("readonly", true);
+        }
         return "restriction_runs";
     }
 
     @RequestMapping(value = "", params="templateId", produces = MediaType.TEXT_HTML_VALUE, method = RequestMethod.GET)
     public String showRestrictionRunsForTemplate(Model model, @RequestParam("templateId") String templateId) {
+        if (readOnly) {
+            model.addAttribute("readonly", true);
+        }
         List<RestrictionRunDocument> runDocuments = restrictionService.findByTemplateId(templateId, new Sort(new Sort.Order(Sort.Direction.ASC, "lastUpdated")));
         model.addAttribute("all_restriction_runs", runDocuments);
         return "restriction_runs";
@@ -48,6 +58,11 @@ public class RestrictionController {
 
     @RequestMapping(value = "/{runid}/delete", produces = MediaType.TEXT_HTML_VALUE, method = RequestMethod.GET)
     public String deleteRun(Model model,@PathVariable String runid,final RedirectAttributes redirectAttributes) {
+
+        if (readOnly) {
+            redirectAttributes.addFlashAttribute("error", "Can't delete, this is a read only version");
+            return "redirect:/restrictions";
+        }
         restrictionService.deleteRun(runid);
         redirectAttributes.addFlashAttribute("message", "Removed run with id: " + runid);
         return "redirect:/restrictions";
@@ -60,6 +75,10 @@ public class RestrictionController {
             @PathVariable String runid,
             final RedirectAttributes redirectAttributes) {
 
+        if (readOnly) {
+            redirectAttributes.addFlashAttribute("error", "Can't run, this is a read only version");
+            return "redirect:/restrictions";
+        }
         RestrictionRunDocument runDocument = restrictionService.findOne (runid);
 
         runDocument = restrictionService.run(runDocument);
